+++ tests/functional_tests/test_general_formatting.py
@@ -60,6 +60,8 @@ from tests.util import check_formatting, read_test_data
         "unformatted/404_create_function_pg_examples.sql",
         "unformatted/405_create_function_snowflake_examples.sql",
         "unformatted/406_create_function_bq_examples.sql",
+        "unformatted/407_alter_function_pg_examples.sql",
+        "unformatted/408_alter_function_snowflake_examples.sql",
         "unformatted/410_create_warehouse.sql",
     ],
 )